lgica de versos

+++ components/Prayers.js
@@ -1,6 +1,7 @@
 import React, { useContext } from "react";
+import Swipeable from "react-native-gesture-handler/ReanimatedSwipeable";
 import { AppContext } from "@/context/contex";
-import { VirtualizedList, Text } from "react-native";
+import { VirtualizedList, Text, Alert } from "react-native";
 import FilterMenu from "./FilterMenu";
 
 import { VStack } from "./ui/vstack";
@@ -8,8 +9,6 @@ import { HStack } from "./ui/hstack";
 import { Card } from "./ui/card";
 import { Switch } from "./ui/switch";
 import { Icon } from "./ui/icon";
-
-import { Filter, ListFilter } from "lucide-react-native";
 
 function Prayers() {
   const {
@@ -22,47 +21,61 @@ function Prayers() {
     getDataFilteredSql,
   } = useContext(AppContext);
 
-  const loadData = async (filter) => {
+  const loadData = async (filter?: any) => {
     const newData = await getDataSql();
-    filter ? setData(filter) : setData(newData)
-    
+    filter ? setData(filter) : setData(newData);
   };
 
-  const handleSetAnswered = async (answered, id) => {
+  const handleSetAnswered = async (answered: number, id: number) => {
     await setAnsweredSql(id, answered);
     await loadData();
   };
 
-  const handleDeletePrayer = async (id) => {
-    await deletePrayerSql(id);
+  const handleDeletePrayer = (id: number) => {
+    Alert.alert(
+      "Eliminar oración",
+      "¿Estás seguro de que deseas eliminar esta oración?",
+      [
+        {
+          text: "Cancelar",
+          style: "cancel",
+        },
+        {
+          text: "Eliminar",
+          style: "destructive",
+          onPress: async () => {
+            await deletePrayerSql(id);
+            await loadData();
+          },
+        },
+      ]
+    );
+  };
+
+  const handleFilterAllAnswered = async () => {
     await loadData();
   };
 
-  const handleFilterAllAnswered  =  async () => {
-    await loadData()
-  };
-  
   const handleFilterPerAnswered = async () => {
-    const data  =  await  getDataFilteredSql(true)
-     setData(data)
-     await loadData(data)
-  }
+    const data = await getDataFilteredSql(true);
+    setData(data);
+  };
 
   const handleFilterNoAnswered = async () => {
-    const data  =  await  getDataFilteredSql(false)
-     setData(data)
-     await loadData(data)
-  }
+    const data = await getDataFilteredSql(false);
+    setData(data);
+  };
 
   return (
     <VStack className="px-5 flex-1 py-5 my-10">
       <HStack className="justify-end p-3">
-        <FilterMenu 
-        handleFilterPerAnswered={handleFilterPerAnswered} 
-        handleFilterAllAnswered={handleFilterAllAnswered}
-        handleFilterNoAnswered={handleFilterNoAnswered}
+        <FilterMenu
+          handleFilterPerAnswered={handleFilterPerAnswered}
+          handleFilterAllAnswered={handleFilterAllAnswered}
+          handleFilterNoAnswered={handleFilterNoAnswered}
         />
       </HStack>
+
       <VirtualizedList
         data={data}
         initialNumToRender={4}
@@ -70,71 +83,88 @@ function Prayers() {
         getItem={(data, index) => data[index]}
         getItemCount={(data) => data?.length ?? 0}
         ListEmptyComponent={() => (
-          <Text
-            style={{
-              color: "black",
-              textAlign: "center",
-              marginTop: 300,
-              fontSize: 18,
-              flex: 1,
-            }}
-          >
-            Escribe tu oración 🙏
-          </Text>
+          <VStack className="mt-50 flex flex-col justify-center items-center">
+            <Text
+              style={{
+                color: "black",
+                textAlign: "center",
+                fontSize: 18,
+              }}
+            >
+              Escribe tu oración 🙏
+            </Text>
+
+          </VStack>
         )}
         renderItem={({ item }) => {
-
           const IconComponent = icons[item.iconName];
 
-          return (
+          const renderRightActions = () => (
             <Card
-              className="w-full h-auto my-2"
-              onPress={() => handleDeletePrayer(item.id)}
-              key={item.id}
+              className="bg-red-600 justify-center items-center"
+              style={{
+                width: 100,
+                marginVertical: 8,
+              }}
+              onTouchEnd={() => handleDeletePrayer(item.id)}
             >
-              <HStack className="gap-2 flex items-center">
-                <Text className="text-white">
-                  Me siento: {item.feeling}
-                </Text>
-
-                <Icon
-                  as={IconComponent}
-                  size={20}
-                  color={item.color}
-                  className="mt-2"
-                />
-              </HStack>
-
-              <HStack>
-                <Text style={{ color: item.color }}>
-                  {item.date}
-                </Text>
-              </HStack>
-
-              <HStack className="justify-between items-center">
-                <Text className="text-white flex-1 mr-3">
-                  {item.prayer}
-                </Text>
-
-                <Switch
-                  size="md"
-                  trackColor={{
-                    false: "#d4d4d4",
-                    true: item.color,
-                  }}
-                  thumbColor="#fafafa"
-                  activeThumbColor="#fafafa"
-                  ios_backgroundColor="#d4d4d4"
-                  value={item.answered === 1}
-                  onToggle={() =>
-                    handleSetAnswered(
-                      item.answered === 1 ? 0 : 1,
-                      item.id
-                    )
-                  }
-                />
-              </HStack>
+              <Text className="text-white font-bold">
+               Eliminar
+              </Text>
             </Card>
+          );
+
+          return (
+            <Swipeable
+              renderRightActions={renderRightActions}
+              rightThreshold={60}
+              overshootRight={false}
+            >
+              <Card className="w-full h-auto my-2">
+                <HStack className="gap-2 flex items-center">
+                  <Text className="text-white">
+                    Me siento: {item.feeling}
+                  </Text>
+
+                  <Icon
+                    as={IconComponent}
+                    size={20}
+                    color={item.color}
+                    className="mt-2"
+                  />
+                </HStack>
+
+                <HStack>
+                  <Text style={{ color: item.color }}>
+                    {item.date}
+                  </Text>
+                </HStack>
+
+                <HStack className="justify-between items-center">
+                  <Text className="text-white flex-1 mr-3">
+                    {item.prayer}
+                  </Text>
+
+                  <Switch
+                    size="md"
+                    trackColor={{
+                      false: "#d4d4d4",
+                      true: item.color,
+                    }}
+                    thumbColor="#fafafa"
+                    activeThumbColor="#fafafa"
+                    ios_backgroundColor="#d4d4d4"
+                    value={item.answered === 1}
+                    onToggle={() =>
+                      handleSetAnswered(
+                        item.answered === 1 ? 0 : 1,
+                        item.id
+                      )
+                    }
+                  />
+                </HStack>
+              </Card>
+            </Swipeable>
           );
         }}
       />
